refactor(houseclean): split js plugin, parser and generator into different folders
