+++ crates/rspack_plugin_javascript/src/lib.rs
@@ -8,6 +8,7 @@ pub(crate) mod dependency;
 mod plugin;
 pub use plugin::*;
 mod ast;
+pub(crate) mod parser_and_generator;
 pub mod runtime;
 pub mod utils;
 pub mod visitors;
@@ -17,25 +18,8 @@ use swc_ecma_minifier::option::{
   terser::{TerserCompressorOptions, TerserEcmaVersion},
   MangleOptions,
 };
-// static JS_HELPERS: Lazy<Helpers> = Lazy::new(Helpers::default);
 
-// use typemap::{Key, TypeMap};
-
-// pub struct JsAst;
-
-// #[derive(Debug)]
-// pub struct Value(swc_ecma_ast::Program);
-
-// impl Key for JsAst {
-//   type Value = Value;
-// }
-
-// fn transform(mut ctx: TypeMap) {
-//   // let mut map = TypeMap::new();
-//   ctx.insert::<JsAst>(Value(swc_ecma_ast::Program::Module(
-//     swc_ecma_ast::Module::dummy(),
-//   )));
-// }
+pub use crate::plugin::infer_async_modules_plugin::InferAsyncModulesPlugin;
 
 #[derive(Debug)]
 pub struct TransformOutput {
